requirements.txt
streamlit
pdfminer.six
chromadb
httpx
docx2txt

streamlit
pdfminer.six
python-pptx
docx2txt
langchain
langchain-openai
httpx
PyMuPDF
# LangChain ecosystem
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
streamlit
pdfminer.six
python-pptx
docx2txt
langchain
langchain-openai
httpx
PyMuPDF
streamlit
pdfminer.six
python-pptx
docx2txt
langchain
langchain-openai
httpx
PyMuPDF

